moved emptyBlock helper function to block-builder within /store
