+++ src/store/block-builder.js
@@ -1,0 +1,18 @@
+/**
+ * @format
+ * @flow
+ */
+import { BlockType } from './';
+
+const emptyBlock: BlockType = {
+	isValid: true,
+	attributes: {
+		content: 'this is a new block',
+	},
+	innerBlocks: [],
+	focused: false,
+};
+
+export function buildEmptyBlock(uid, name) {
+	return {uid: uid, name: name, ...emptyBlock}
+}
